ハンドラー関数をカスタムフックに切り出し
- useAnalytics: アナリティクス関連のstate/handlers
  - analyticsData, unblockHistory
  - handleReblock, handleResetAnalytics, handleStopTracking
  - handleRefreshAnalytics, handleAddSiteToTrack
- useLicense: ライセンス関連のhandlers
  - handleUpgrade, handleManageSubscription
- options.tsxから約140行削減

Closes #51
