+++ src/hooks/index.ts
@@ -1,4 +1,6 @@
+export * from './useAnalytics';
 export * from './useBlocklist';
+export * from './useLicense';
 export * from './useSchedules';
 export * from './usePresets';
 export * from './useBackgroundStats';
